requirements.txt
pandas
pypdf
python-pptx
google.generativeai
